Remove AsioHeaders dependency (using dev branch of ipaddress)
