+++ DESCRIPTION
@@ -33,11 +33,12 @@ Suggests:
     tidyverse,
     vdiffr
 LinkingTo: 
-    AsioHeaders,
     ipaddress,
     Rcpp
 VignetteBuilder: 
     knitr
+Remotes:
+    davidchall/ipaddress@reduce-asio-dependence
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
