--- src/Makevars
@@ -1,1 +1,0 @@
-PKG_CXXFLAGS = -DASIO_STANDALONE -DASIO_NO_DEPRECATED
